git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@821 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/G3D.lib/include/G3D/Table.h
@@ -380,7 +380,7 @@ public:
        Returns the average size of non-empty buckets.
     */
     float debugGetAverageBucketSize() const {
-        size_t num = 0;
+        uint64 num = 0;
 
         for (size_t b = 0; b < m_numBuckets; b++) {
             Node* node = m_bucket[b];
